web-platform_22_10_dev/web/requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Deprecated==1.2.13
Django==4.1.2
django-cors-headers==3.13.0
django-environ==0.9.0
django-grappelli==3.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
frozenlist==1.3.1
gunicorn==20.1.0
httplib2==0.20.4
idna==3.4
kombu==5.2.4
lxml==4.9.1
multidict==6.0.2
openpyxl==3.0.10
packaging==21.3
Pillow==9.2.0
prompt-toolkit==3.0.31
psycopg2==2.9.5
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyodbc==4.0.34
pyparsing==3.0.9
pytz==2022.5
redis==4.3.4
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.5
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
yarl==1.8.1
